+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+### v1.1
+- Add configuration option for scaling the text
+- Add configuration option for displaying a shadow under the text
+
 ### v1.0
 - Initial release